move order summery into its own file
